verbs: Use netlink to discover uverbs devices instead of sysfs
The netlink query gives us the ibdev idx, which is mostly unique for the
device, and serves as a stable id across device renames. This makes verbs
much more robust if the device is renamed during the operation of a verbs
user.

Also, netlink only returns devices which are actually visible in the net
namespace of the process, simplifing discovery.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>