efa: Fill send operation type in TX descriptor
The send operation type was mistakenly not filled in the TX meta descriptor
and resulted in the provider always passing invalid op (0),
which the device wrongly treated as a send operation.

To fix the bug in a backwards compatible way, we change zero to represent
send operation and remove the invalid operation.
While at it, the unsupported RDMA types are removed as well.

Signed-off-by: Daniel Kranzdorf <dkkranzd@amazon.com>
Reviewed-by: Yossi Leybovich <sleybo@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>