libhns: Modify pi vlaue when cq overflows
When exiting "for loop" in __hns_roce_v2_cq_clean(), the actual value of pi
will be increased by 1, which is compatible with the next calculation.
But when pi is equal to "ci + hr_cq-> ib_cq.cqe", the "break"
was called and the pi is actual value, it will lead one cqe
still existing, so the "==" should be modify to ">".

Signed-off-by: Yangyang Li <liyangyang20@huawei.com>